+++ packages/extensions/git/src/__tests__/ChangesTab.commit.test.tsx
@@ -22,6 +22,7 @@ function renderTab() {
       onShowOutput={() => {}}
       fileMaskEnabled={false}
       fileMaskInput=""
+      refreshToken={0}
     />,
   );
 }